requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
flask==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
werkzeug==3.1.3
ollama
time
matplotlib.pyplot